bug: fix sharing violation exception on windoes.
After uploading a file and try to move it to backup folder, it will
trigger sharing violation on windows.

File lock is holded even it is done on windows. Uploaded file is
locked by Google Drive API, but there is no function to close file
handle in API actively.

The file handle will be released after exting file uploading function.
To avoid this exception, use a wrapper function to handle backup action.